requirements.txt
pyserial==3.5
opencv-python==4.6.0.66
numpy==1.23.4